+++ G07P3/src/src/bloating/IBloating.java
@@ -6,5 +6,5 @@ import java.util.Random;
 import src.cromosoma.Cromosoma;
 
 public interface IBloating {
-	public void control(int profundidadMedia, int mediaFitness, ArrayList<Cromosoma> poblacion, Random rand, double probMuerte, int actGen, int numGen);
+	public void control(int profundidadMedia, double mediaFitness, ArrayList<Cromosoma> poblacion, Random rand, double probMuerte, int actGen, int numGen);
 }
